Nr/stock counts (#58)
* *FEAT* se añade migracion, controllers y test para el crud de stockcounts

* *FIX* se corrige una validación para la petición del campo cantidad quitando una letra que no debería de ir

* *FEAT* se añaden comentarios para los test

* *FIX* se añaden cambios requeridos eliminando comentarios, throw y validando siempre los procesos

* *FEAT* se añaden las rutas al archivo de permisos

* *FEAT* se añaden cambios para los test añadiendo seeds de usuarios y roles

* *FEAT* se agregan cambios en los controladores de stock counts para obtener el detalle por stock_counts y no stock_counts_detail

* *FEAT* se agrega la correcta relacion en la obtencion del detalle de conteos

* *FIX* se corrige la validacion del estado al actualizar conteos

* *FIX* correccion al metodo de validacion

+++ app/Http/Modules/StockCounts/StockCountsController.php
@@ -26,21 +26,24 @@ class StockCountsController extends Controller
 
             $stockCount = StockCounts::create($request->validated());
 
-            $stockCountDetailProduct = $request->stock_counts_detail_product;
-            $stockCountDetailQuantity = $request->stock_counts_detail_quantity;
+            if(!empty($request->stock_counts_detail_product)) {
+                $stockCountDetailProduct = $request->stock_counts_detail_product;
+                $stockCountDetailQuantity = $request->stock_counts_detail_quantity;
 
-            $arrDetail = array();
+                $arrDetail = array();
 
-            for($int = 0; $int < count($stockCountDetailProduct); $int++) {
-                $countsDetail = array(
-                    'stock_count_id' => $stockCount->id,
-                    'product_id' => $stockCountDetailProduct[$int],
-                    'quantity' => $stockCountDetailQuantity[$int],
-                );
-                $arrDetail[] = $countsDetail;
+                for($int = 0; $int < count($stockCountDetailProduct); $int++) {
+                    $countsDetail = array(
+                        'stock_count_id' => $stockCount->id,
+                        'product_id' => $stockCountDetailProduct[$int],
+                        'quantity' => $stockCountDetailQuantity[$int],
+                    );
+                    $arrDetail[] = $countsDetail;
+                }
+    
+                StockCountsDetail::create($arrDetail);
             }
-
-            StockCountsDetail::create($arrDetail);
+            
 
             DB::commit();
             return $this->showOne($stockCount, 201);
@@ -60,27 +63,27 @@ class StockCountsController extends Controller
 
             $stockCount->update($request->validated());
             
-            $stockCountDetailProduct = $request->stock_counts_detail_product;
-            $stockCountDetailQuantity = $request->stock_counts_detail_quantity;
+            if(!empty($request->stock_counts_detail_product)) {
+                $stockCountDetailProduct = $request->stock_counts_detail_product;
+                $stockCountDetailQuantity = $request->stock_counts_detail_quantity;
 
-            $arrDetail = array();
+                $arrDetail = array();
 
-            for($int = 0; $int < count($stockCountDetailProduct); $int++) {
-                $countsDetail = array(
-                    'stock_count_id' => $stockCount->id,
-                    'product_id' => $stockCountDetailProduct[$int],
-                    'quantity' => $stockCountDetailQuantity[$int],
-                );
-                $arrDetail[] = $countsDetail;
-            }
+                for($int = 0; $int < count($stockCountDetailProduct); $int++) {
+                    $countsDetail = array(
+                        'stock_count_id' => $stockCount->id,
+                        'product_id' => $stockCountDetailProduct[$int],
+                        'quantity' => $stockCountDetailQuantity[$int],
+                    );
+                    $arrDetail[] = $countsDetail;
+                }
 
-            StockCountsDetail::where('stock_count_id', '=', $stockCount->id)->delete();
-            
-            foreach($arrDetail AS $key => $value) {
-                //StockCountsDetail::update($value);
-                StockCountsDetail::create($value);
-            }
-            
+                StockCountsDetail::where('stock_count_id', '=', $stockCount->id)->delete();
+                
+                foreach($arrDetail AS $key => $value) {
+                    StockCountsDetail::create($value);
+                }   
+            }            
 
             DB::commit();
             return $this->showOne($stockCount);
